Move the server creation out of the application
This means that the application only really depends on the server
factory, and all it does is asking the server lib to create a server and
runs it. No more dependencies on ingredients in the application as it
doesn't know what kind of server it's running and creation of the server
is internal to the server lib.