Add Catch2, example tests, fix bugs discovered, and run tests in CI
The Catch2 test framework was chosen because why not? It's a pretty nice
framework.

The bugs discovered and fixed were as follows:
* ingredient needed to be a public dependency of recipe.
* find_by_id in both the RecipeRepository and the IngredientRepository
threw an exception instead of returning std::nullopt when the id was
missing from the database.

The spaceship-operators are not defaulted as it's not yet implemented
for std::string in the version of Clang we're using on CI.