Fix VS2019-build
* Simplify conan boost dependency.
* Make sure range-v3 provides the compilation flags it requires.
* Silence deprecation warnings in yaml-cpp.
* Replace deprecated range/v3/to_container with the new ones.
* Make sure autobahn has its dependencies.
* Remove unreferenced argument to fix warning.

The deprecation warnings in yaml-cpp have been fixed in master, but
they've yet to tag a release with the fix.

range/v3/to_container is deprecated in favour of
range/v3/range/conversion.

autobahn has never correctly pulled in msgpack or websocketpp when
building with conan due to the casing differences in the names between
what they use and what conan uses. How has this ever worked? :D